Building cards.json manually now
The mtgjson project is good, but it's not a good node api. It's a great flat-file api though, so I'm leveraging the combineAllSets again in the gulpfile to get what I need. The reason being that the tests were failing because requesting mtgjson async was too slow and I didn't want to move it outside of the Api class's constructor.